+++ GoldenEleganceProyecto/VistasAngular/src/app/auth/pages/form-reset-password/form-reset-password.component.ts
@@ -1,18 +1,16 @@
 import { AuthService } from './../../../services/auth.service';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { ResetPassword } from 'src/app/models/reset-password.model';
 import { confirmarPasswordValidation } from 'src/app/shared/helpers/confirmar-password.validator';
-import Swal from 'sweetalert2';
-
 @Component({
   selector: 'app-form-reset-password',
   templateUrl: './form-reset-password.component.html',
   styleUrls: ['./form-reset-password.component.css']
 })
 export class FormResetPasswordComponent implements OnInit {
-  constructor(private formB: FormBuilder, private route: Router, private activaterouter: ActivatedRoute, private authS: AuthService) { }
+  constructor(private formB: FormBuilder,  private activaterouter: ActivatedRoute, private authS: AuthService) { }
 
 
   resetPasswordForm!: FormGroup;
@@ -43,27 +41,7 @@ export class FormResetPasswordComponent implements OnInit {
       this.resetPasswordObj.emailToken=this.emailtoken;
 
       this.authS.resetPassword(this.resetPasswordObj)
-      .subscribe({
-        next:()=>{
-          Swal.fire({
-            position: 'top-end',
-            icon: 'success',
-            title: 'Se restablecio correctamente la contraseña',
-            showConfirmButton: false,
-            timer: 1500,
-          });
-          this.route.navigate(['login']);
-        },
-        error:()=>{
-          Swal.fire({
-            position: 'top-end',
-            icon: 'warning',
-            title: 'Sucedio un error al cambiar la contraseña',
-            showConfirmButton: false,
-            timer: 1500,
-          });
-        }
-      })
+     
     }
   }
 }
