+++ GoldenEleganceProyecto/VistasAngular/src/app/auth/pages/login/login.component.ts
@@ -4,7 +4,6 @@ import { AuthService } from 'src/app/services/auth.service';
 import Swal from 'sweetalert2';
 import { Router } from '@angular/router';
 import { IResponseToken } from '../../../models/IResponseToken.interface';
-import { IResponse } from 'src/app/models/IResponse.interface';
 import { catchError, of } from 'rxjs';
 import { HttpErrorResponse } from '@angular/common/http';
 import { UserStoreService } from 'src/app/services/Usuarios/user-store.service';
@@ -21,8 +20,8 @@ export class LoginComponent {
     private route: Router,
     private userStoreS: UserStoreService
   ) { }
-  role= "";
-  claseDiv= 'container';
+  role = "";
+  claseDiv = 'container';
 
   sigUpForm: FormGroup = this.fb.group({
     Nombre: ['', Validators.required],
@@ -47,111 +46,12 @@ export class LoginComponent {
 
   onLogin() {
     if (this.loginForm.valid) {
-      this.auth
-        .iniciarSesion(this.loginForm.value)
-        .pipe(
-          catchError((error: HttpErrorResponse) => {
-            let response: IResponseToken = {
-              success: false,
-              helperData: "",
-              message: "",
-              token:"",
-              refreshToken:""
-            };
-            if (error.status == 404 || error.status == 400 || error.status == 401) {
-              if (error.error != null) {
-                response = error.error;
-              }
-            }
-            return of(response);
-          })
-        )
-        .subscribe((response) => {
-          if (response.success) {
-            Swal.fire({
-              position: 'top-end',
-              icon: 'success',
-              title: 'Inicio de sesión correcto',
-              showConfirmButton: false,
-              timer: 2500,
-            });
-            const rol = this.StorageUser(response);
-            if(Number(rol) == 1 ){
-              this.route.navigate(['base/dashboard']);
-
-            }else{
-              this.route.navigate(['base/inicio']);
-
-            }
-          } else {
-            Swal.fire({
-              position: 'top-end',
-              icon: 'warning',
-              title: 'Error al iniciar sesión',
-              text: response.message,
-              timer: 2500,
-
-            });
-          }
-        });
+      this.auth.iniciarSesion(this.loginForm.value)
     }
   }
-  private StorageUser(response: IResponseToken) {
-    this.auth.almacenarToken(response.token);
-    this.auth.almacenarRefreshToken(response.refreshToken);
-    const tokenPayload = this.auth.decodedToken();
-    this.userStoreS.setUsernameFromStore(tokenPayload.name);
-    this.userStoreS.setRoleFromStore(tokenPayload.role);
-    this.userStoreS.getRoleFromStore()
-      .subscribe(valor => {
-        const rolFromToken = this.auth.getRolFromToken();
-        this.role = valor || rolFromToken;
-      })
-
-      return this.role;
-  }
-
   onSingUp() {
     if (this.sigUpForm.valid) {
-      this.auth
-        .registrarse(this.sigUpForm.value)
-        .pipe(
-          catchError((error: HttpErrorResponse) => {
-            let response: IResponse = {
-              success: false,
-              helperData: 'Datos incorrectos',
-              message: 'No se pudo iniciar sesion',
-            };
-            if (error.status == 404 || error.status == 400 || error.status == 401) {
-              if (error.error != null) {
-                response = error.error;
-              }
-            }
-            return of(response);
-          })
-        )
-        .subscribe((response: IResponse) => {
-          if (response.success) {
-            Swal.fire({
-              position: 'top-end',
-              icon: 'success',
-              title: 'Usuario registrado correctamente',
-              showConfirmButton: false,
-              timer: 1500,
-            });
-            this.sigUpForm.reset();
-            this.route.navigate(['login']);
-          } else {
-            Swal.fire({
-              position: 'top-end',
-              icon: 'warning',
-              title: 'Usuario no registrado',
-              text: response.message,
-              showConfirmButton: false,
-              timer: 3500,
-            });
-          }
-        });
+       this.auth.registrarse(this.sigUpForm.value);
     }
   }
 }
