+++ GoldenEleganceProyecto/VistasAngular/src/app/auth/pages/reset-password/reset-password.component.ts
@@ -19,26 +19,6 @@ export class ResetPasswordComponent {
  
   onSubmit(){
   this.authS.sendResetPassword(this.emailPassword.controls['Correo'].value)
-  .subscribe({
-    next:(resp)=>{
-      Swal.fire({
-        position: 'top-end',
-        icon: 'success',
-        title: 'Se realizo correctamente el envio del email' + resp.message,
-        showConfirmButton: false,
-        timer: 1500,
-      });
-    }, 
-    error:(error)=>{
-      Swal.fire({
-        position: 'top-end',
-        icon: 'warning',
-        title: 'No se pudo enviar' + error,
-        showConfirmButton: false,
-        timer: 1500,
-      });
-    }
-  })
   }
 }
