+++ GoldenEleganceProyecto/VistasAngular/src/app/base-admin/pages/inicio/inicio.component.html
@@ -4,7 +4,7 @@
     <div  *ngFor="let item of productos" >
         <div class="cards">
             <figure>
-                <img src="../../../../assets/img/imag.png">
+                <img src="../../../../assets/img/imag.png" alt="hi">
             </figure>
             <div class="contenido-cards">
                 <h3>{{item.nombreProducto }}</h3>
