+++ GoldenEleganceProyecto/appsettings.json
@@ -13,6 +13,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": " Data Source =DESKTOP-RABRFD1\\MSSQLSERVER01 ; initial catalog = DBGoldenElegance; Integrated Security = True;"
+    "DefaultConnection": " Data Source =BRYANXG ; initial catalog = DBGoldenElegance; Integrated Security = True;"
   }
 }
